requirements.txt
colorama==0.4.6
discord
Flask==3.0.2
httpx
licensing==0.43
requests==2.31.0
tls-client==1.0.1
typing_extensions==4.10.0
pymongo
fastapi
websocket-client
uvicorn
websockets
toml
pystyle
discord_webhook
pytz
PyYAML